Auditallow apk_data_file execute
For libraries, apps should be migrating to the more modern approach of
storing them in the apk uncompressed and mapping them directly from it.

This is the most modern approach available for executables and is better
than using app data, but ideally it wouldn't be done. For now, audit use
of `execute_no_trans` anyway while this is given more thought.

Signed-off-by: anupritaisno1 <www.anuprita804@gmail.com>
Change-Id: I788e1249ff480f0956f2742cf962129d0869f535